First controller - but still an error
The Pluralsight tutorial uses a function in the global scope as the
controller. However, an overlay on the video indicates that this
structure is no longer allowed beginning, I believe with Angular 1.2.5.
This StackOverflow post,
http://stackoverflow.com/questions/25111831/controller-not-a-function-got-undefined-while-defining-controllers-globally,
indicates that this construct was not allowed as of Angular 1.3.

The tutorial indicates that Scott Allen will describe later how to
expose the controller function is a module using an IIFE; however, the
StackOverflow post also describes how to register the function. For now,
I will use the StackOverflow approach. (Based on previous tutorials, I
believe this is the "accepted" approach.)