One can use an anonymous function
Previous commits defined the `ng-controller` to be a function named
`MainController`. This commit uses an anonymous function, bound to the
name "MainController", as the controller.

Although this works, it requires defining the function inline. In this
particular situation, the body of the function consists of a single
line: an inline function is comprehensible. Although many programmers
define complicated functions inline, I much prefer to define complicated
functions separately. JavaScript, unlike ClojureScript or Lisp, allows
using a function before its definition so that I can define the function
later if it is not critical to understand when reading the code
top-to-bottom.

This style allows me to read the code linearly instead of reading an
inline definition and needing to say, "Push this block of code onto your
stack for later retrieval."