Bind object to Angular controller
This commit exposes both a single value, `message`, and a complex value
(a JavaScript object), `person`, to the controller.

The controller then uses this object to not only bind textual values,
`salutation` and `lastName`, but to bind to attribute values of an `img`
tag.

Rendering the page produces an error because when the page first
renders, Angular is not fully loaded. However, I will fix this issue in
the next commit.