Change calls to Angular to address minification
Most production JavaScript code undergoes minification before being
transmitted over the wire. One typical result of this minification is
changing the names of variables from meaningful names like `$scope` to
meaningless but shorter names like `n`.

Angular, when invoking client code like controllers, looks for the
meaningful names (`$scope`) in the parameter list of the controllers.
When it finds the name `n`, Angular decides **not** to pass `$scope` as
an actual controller argument. The application then fails.

To address this issue, Angular provides a registration syntax for
controllers that is robust in the face of minification. This syntax
involves an array as the second argument to `controller`. The array
contains the string names of arguments, in the correct order, that
Angular must pass to the controller invocation even if it **did not**
find meaningful parameter names; that is, `$scope`.